### 2023-12-04 #6119 @editorialbot
[PRE REVIEW]: MRdataset: A unified and user-friendly interface to medical imaging datasets
**Submitting author:** @sinhaharsh (<a href="http://orcid.org/0000-0002-8498-8559">Harsh Sinha</a>)
**Repository:** https://github.com/Open-Minds-Lab/MRdataset
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.33
