### 2023-12-04 #6119 @Kevin-Mattheus-Moerman
@sinhaharsh thanks for this submission. I am the AEiC on this track and here to help with initial steps. I have just added the `query-scope` label here. This is to trigger a scope review by the editorial board. I have done this because I need some help to determine if this work is in scope. In particular some domain experts should weigh in to see if the scientific research functionality provided here is significant enough to warrant a JOSS publication.
The scope review should take about 2 weeks to complete.
