### 2023-12-05 8539a34
add figure to paper.md
