### 2023-12-29 #6119 @raamana
Hi @Kevin-Mattheus-Moerman, did the query for scope get triggered properly?

I am quite biased as a developer and an author, I can't see why our paper would be out of scope, and I'd think any active informatics researchers neuroscience world would likely agree with us too :)
