### 2024-01-11 #6119 @raamana
Hi @Kevin-Mattheus-Moerman , can you update us on whats happening? its been over 5 weeks now (although I do understand there were some holiday breaks etc). if we don't hear from you in the next week or two (which would disappoint me greatly as someone who published 4 papers in JOSS before), we would have to explore alternative venues.
