### 2024-01-11 #6119 @Kevin-Mattheus-Moerman
@raamana @sinhaharsh apologies for the delay. With the holidays, and the start of semester for many, this scope review has taken longer than expected. Having said that I should also apologize for not responding/providing and update sooner.
Based on the scope review from the editorial board so far I believe we can proceed and say this work may be in scope.
I will remove the query scope label now, and will start by inviting an editor. One concern was that this work may not offer more than a wrapper around pydicom. It may be helpful if you responded here with information around how your work goes beyond that.
