### 2024-01-12 #6119 @raamana
Thanks for the response - as out of domain editor, I appreciate your confusion and questions, and we hope we have sufficiently answered them in short paper we submitted for review: https://github.com/Open-Minds-Lab/MRdataset/blob/joss/joss/paper.pdf

MRDataset was developed to reduce the complexity in dealing with large variations in **dataset** formats! I repeat, it is NOT file format (like DICOM), but a DATASET format like BIDS and other more complex and esoteric structures in use in the neuroscience world. The Base classes (such as `BaseMRDataset`) and the derived classes we designed ( such as `DicomDataset`) are aimed to offer a uniform interface to an user without expertise in dealing with variations in these DATASET formats (again, not file formats). Other than pybids (which focuses exclusively on BIDS), we are not aware of other libraries doing this (we'd be happy to cover other libraries if we missed them). Perhaps we didn't communicate that well enough in our paper and we hope to revise it during the review process.

Besides the unified interface at the dataset level, though physics-based representation of individual MRI physics parameters (via [`protocol`](https://www.github.com/raamana/protocol) ), we offer deep cross-vendor conversion (across Siemens, GE, Philips etc) for MRI physics parameter units and values, that is essential for the protocol compliance research based on over 20 open datasets that MRdataset enables.

PS: also, regarding the concern that "this work may not offer more than a wrapper around pydicom", the same accusation can be laid against pybids and imagedata (and even nibabel), but talk to any of the neuroinformatics researchers, they would make it clear how useful and critical these libraries are.
