### 2024-01-15 #6119 @Kevin-Mattheus-Moerman
@raamana thanks for getting back to me on that. My domain is indeed different (However, I have worked on MRI sequences, and did in the past implement a DICOM/analyze importer). The comments I posted originated from points raised by other editorial board members who are closer aligned with this domain. Either way, we are passed this now. I think the information you provide is helpful. We are now waiting for this to receive a handling editor. I'll ping the one I invited now again.
