### 2024-01-15 #6119 @emdupre
Hi @Kevin-Mattheus-Moerman, thanks for the invitation ! I needed to re-check the JOSS COI policy on for both of the newly invited submissions.

For this one, I have a few authorships with the senior author, [the most recent ](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC10516110/) having been accepted this month, though [this 2021 paper ](https://pubmed.ncbi.nlm.nih.gov/34414422/)would also qualify under [the COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). As these were large efforts in which Dr Raamana and I had minimal interaction, however, I believe that I can review this submission impartially. I'll leave, though, the final decision up to you !
