### 2024-01-16 #6119 @Kevin-Mattheus-Moerman
@emdupre those would quality as potential/perceived COIs so I'll have to recruit an alternative editor. Thanks for reporting that.
