### 2024-01-18 #6119 @raamana
Thank you Marcel for accepting and helping facilitate the review! I will try to come up with some names, but I must admit most of the names that come to my mind right now (those who'd appreciate what this has to offer) are folks (like @emdupre ) I've collaborated with in some form over big community projects or small hackathon projects

great reviewers would be potential future users, those **that actually have reasonable experience in acquiring MRI scans on physical scanners in the real-world** (not just processing them after they shared in BIDS format etc) and know the issues with managing DICOM data
