### 2024-01-23 #6119 @raamana
Hi Marcel and Kevin, back in the day when JOSS started, there was an excel of volunteers willing to review. Does JOSS still maintain it? also, do you have any tools for automatic checks of COI?
