### 2024-01-23 #6119 @mstimberg
> Hi Marcel and Kevin, back in the day when JOSS started, there was an excel of volunteers willing to review. Does JOSS still maintain it? also, do you have any tools for automatic checks of COI?

The old sheet has been replaced by this system here: https://reviewers.joss.theoj.org/reviewers AFAIK, we don't have any automatic system for checking COIs.

I found two reviewers that agreed to review, though. I will assign them now and we can start the review process :rocket:
