### 2024-01-23 #6269 @mstimberg
👋🏼  @sinhaharsh, @raamana, @htwangtw, @djmannion, this is the review thread for the paper. All of our communications will happen here from now on. Looking forward to an enjoyable and constructive review :blush:

at the top of a new comment in this thread.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#6269` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them, instead of waiting until you've reviewed the entire package.
There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/6269#issue-2096460730).
