### 2024-01-23 #6269 @raamana
thanks again everyone!

if i am reading the [table](https://github.com/openjournals/joss-reviews/issues/6269#issuecomment-1906422875) here correctly, there is a lot of blank files, which is kinda strange. we will take a look and purge any unnecessary ones (if not needed by library or docs)
