### 2024-01-23 #6269 @raamana
unless it is number of blank lines? 🤣
