### 2024-01-24 9fe32f8
Documentation : Add description and change figure
