### 2024-01-24 Open-Minds-Lab/MRdataset#24 @htwangtw
[JOSS] `protocol` version when installing from source vs install through pip 
openjournals/joss-reviews#6269

* MRdataset version: 0.1.dev
* Python version: 3.10.12
* Operating System: Ubuntu 22.04.3 LTS

### Description

This is part of the JOSS submission review.

In the installation instruction, the stable release will install a key dependency `protocol` version 0.3, however if we strictly follow the install from source instruction, it will be 0.1.

These two things should be consistent.

### What I Did
```
virtualenv env-stable
source env-stable/bin/activate
pip install -U MRdataset
pip list
```
From source:
```
git clone git://github.com/Open-Minds-Lab/MRdataset
cd MRdataset
virtualenv env
source env/bin/activate
pip install git+https://github.com/sinhaharsh/protocol.git#egg=protocol
pip install -r requirements_dev.txt
pip install -e .
pip list
```
