### 2024-01-24 Open-Minds-Lab/MRdataset#24 @raamana
that's a great catch Hao, thank you. I guess we never attempted install from source :)

we will update the docs to make it they should use `pip install -U MRdataset` when possible, and only when they trying to contribute as a developer should they use the other instructions.

I personally don't understand the parameter `#egg=protocol` in this command
in `pip install git+https://github.com/sinhaharsh/protocol.git#egg=protocol`
but my guess is we don't need it. @sinhaharsh and I take a look and fix the docs asap.
