### 2024-01-24 Open-Minds-Lab/MRdataset#25 @htwangtw
[JOSS] BIDS option not implemented
openjournals/joss-reviews#6269

* MRdataset version: 0.1.dev
* Python version: 3.10.12
* Operating System: Ubuntu 22.04.3 LTS

### Description

I modify the jupyter notebook example and found the BIDS option is flagged as not implemented.
The documentation and the paper gave me the impression that at least the dicom and nifti options are available. This feature has to be added to fit the documentation, and the scope of the paper and the software should be clarified.

### What I Did

```
from pathlib import Path
from MRdataset import import_dataset
import zipfile

MRdataset_root = Path(__name__).resolve().parents[1]
DATA_ARCHIVE = MRdataset_root / 'examples'/ 'example_bids_dataset.zip'
DATA_ROOT = Path('/tmp/')
with zipfile.ZipFile(DATA_ARCHIVE, 'r') as zip_ref:
    zip_ref.extractall(DATA_ROOT)
print((DATA_ROOT/'example_bids_dataset').exists())
dicom_dataset = import_dataset(data_source=DATA_ROOT/'example_bids_dataset',
                               ds_format='bids')
```
Output:
```
File [~/projects/reviews/MRdataset/MRdataset/common.py:130](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:130), in find_dataset_using_ds_format(dataset_ds_format)
    [128](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:128)     dataset_class = DicomDataset
    [129](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:129) else:
--> [130](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:130)     raise NotImplementedError(
    [131](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:131)         f'Dataset ds_format {dataset_ds_format} is not implemented. '
    [132](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:132)         f"Valid choices are {', '.join(VALID_DATASET_FORMATS)}")
    [133](https://file+.vscode-resource.vscode-cdn.net/home/haoting/projects/reviews/MRdataset/docs/~/projects/reviews/MRdataset/MRdataset/common.py:133) return dataset_class

NotImplementedError: Dataset ds_format bids is not implemented. Valid choices are dicom, bids
```
