### 2024-01-24 Open-Minds-Lab/MRdataset#26 @htwangtw
[JOSS] Jupyter notebook linked in documentation is outdated
openjournals/joss-reviews#6269

The tutorial notebook doesn't match the `usage.ipynb` file in the repository.
The online version also includes function that's not present in the current version of the library (such as `print_tree`)
