### 2024-01-25 fd0895a
Add : BidsDataset class
