+++ MRdataset/__init__.py
@@ -16,6 +16,7 @@ from MRdataset.config import MRDS_EXT, DatasetEmptyException
 from MRdataset.dicom_utils import is_dicom_file
 from MRdataset.utils import valid_dirs
 from MRdataset.base import BaseDataset
+from MRdataset.bids import BidsDataset
 from MRdataset.dicom import DicomDataset
 
 try:
