+++ MRdataset/common.py
@@ -4,6 +4,7 @@ from typing import Union, List
 
 from MRdataset import logger
 from MRdataset.base import BaseDataset
+from MRdataset.bids import BidsDataset
 from MRdataset.config import VALID_DATASET_FORMATS
 from MRdataset.dicom import DicomDataset
 from MRdataset.utils import random_name, check_mrds_extension
@@ -126,6 +127,8 @@ def find_dataset_using_ds_format(dataset_ds_format: str):
     # Import the module "{ds_format}_dataset.py"
     if dataset_ds_format == 'dicom':
         dataset_class = DicomDataset
+    elif dataset_ds_format == 'bids':
+        dataset_class = BidsDataset
     else:
         raise NotImplementedError(
             f'Dataset ds_format {dataset_ds_format} is not implemented. '
