### 2024-01-29 a81c2d7
Add code to create a new directory if args.output_dir doesn't exist
