### 2024-01-29 4ee94b7
Create tests for BIDS dataset
