### 2024-01-30 53c2390
Create more tests for BIDS dataset
