### 2024-01-30 a90743a
Enhance BIDS dataset : Documentation and catch errors
