+++ MRdataset/bids.py
@@ -7,26 +7,45 @@ from MRdataset import logger
 from MRdataset.base import BaseDataset
 from MRdataset.config import VALID_BIDS_DATATYPES
 from MRdataset.dicom_utils import is_bids_file
-from MRdataset.utils import folders_with_min_files, valid_dirs
+from MRdataset.utils import folders_with_min_files, valid_dirs, read_json
 
 
 class BidsDataset(BaseDataset, ABC):
-    """Class to represent a BIDS dataset"""
+    """
+    Class to represent a BIDS dataset. It is a subclass of BaseDataset.
+    It gathers data from JSON files.
+
+    Parameters
+    ----------
+    data_source : str or List[str]
+        The path to the dataset.
+    pattern : str
+        The pattern to match for JSON files.
+    name : str
+        The name of the dataset.
+    config_path : str
+        The path to the config file.
+    verbose : bool
+        Whether to print verbose output on console.
+    ds_format : str
+        The format of the dataset. One of ['dicom', 'bids'].
+    """
 
     def __init__(self, data_source, pattern="*.json",
                  name='BidsDataset',
                  config_path=None,
                  verbose=False,
-                 ds_format='bids',
                  output_dir=None,
+                 min_count=1,
                  **kwargs):
-        super().__init__(data_source=data_source, name=name, ds_format=ds_format)
+
+        super().__init__(data_source=data_source, name=name, ds_format='bids')
         self.data_source = valid_dirs(data_source)
         self.pattern = pattern
         self.config_path = config_path
         self.verbose = verbose
         self.config_dict = None
-        self.min_count = 1
+        self.min_count = min_count
 
         try:
             self.output_dir = Path(output_dir)
@@ -36,11 +55,31 @@ class BidsDataset(BaseDataset, ABC):
 
         self.output_dir.mkdir(exist_ok=True, parents=True)
 
+        # read the config file
+        try:
+            self.config_dict = read_json(Path(self.config_path))
+        except (FileNotFoundError, ValueError) as e:
+            logger.error(f'Unable to read config file {self.config_path}')
+            raise e
+
+        self.includes = self.config_dict.get('include_sequence', {})
+        self.include_nifti_headers = self.includes.get('nifti_header', False)
+
     def load(self):
+        """
+        Default method to load the dataset. It iterates over all the folders
+        in the data_source and finds subfolders with at least min_count files
+        matching the pattern. It then processes each subfolder and adds the
+        sequence to the dataset.
+        """
+
         for directory in self.data_source:
+            # find all sub-folders with at least min_count files matching the
+            # pattern
             subfolders = folders_with_min_files(directory, self.pattern,
-                                                 self.min_count)
+                                                self.min_count)
             for folder in subfolders:
+                # process each folder
                 sequences = self._process(folder)
                 for seq in sequences:
                     self.add(subject_id=seq.subject_id,
@@ -49,28 +88,37 @@ class BidsDataset(BaseDataset, ABC):
                              seq_id=seq.name, seq=seq)
 
     def _filter_json_files(self, folder):
+        """Filters the JSON files from the folder."""
         json_files = sorted(folder.glob(self.pattern))
-        valid_bids_files = filter(is_bids_file, json_files)
+        valid_bids_files = list(filter(is_bids_file, json_files))
         if not valid_bids_files:
             logger.info(f'No valid BIDS files found in {folder}')
             return []
         return valid_bids_files
 
     def _process(self, folder):
+        """Processes the folder and returns a list of sequences."""
         json_files = self._filter_json_files(folder)
         sequences = []
         for i, file in enumerate(json_files):
-            seq = BidsImagingSequence(bidsfile=file, path=folder)
+            try:
[…]
