### 2024-01-30 43bea6c
Documentation : Remove protocol from installation.rst Issue : https://github.com/Open-Minds-Lab/MRdataset/issues/24
