### 2024-01-30 Open-Minds-Lab/MRdataset#24 @sinhaharsh
Compiling protocol from source was initially required. It's surely not required anymore. It is now addressed in the documentation. Thank you Hao.
