### 2024-01-31 2de1ea0
Fix: Codacy static analysis
