### 2024-01-31 72d7986
Fix: Codacy static analysis
