+++ MRdataset/tests/test_dicom.py
@@ -17,7 +17,8 @@ from MRdataset.tests.simulate import make_compliant_test_dataset, \
 THIS_DIR = Path(__file__).parent.resolve()
 
 
-@settings(suppress_health_check=[HealthCheck.too_slow], max_examples=50, deadline=None)
+@settings(suppress_health_check=[HealthCheck.too_slow], max_examples=50,
+          deadline=None)
 @given(st.integers(min_value=1, max_value=10),
        st.floats(allow_nan=False,
                  allow_infinity=False),
@@ -32,23 +33,16 @@ def test_parse_compliant_dataset(num_subjects,
                                               repetition_time,
                                               echo_train_length,
                                               flip_angle)
-    mrd = import_dataset(fake_ds_dir, config_path=THIS_DIR / 'resources/mri-config.json',
+    mrd = import_dataset(fake_ds_dir,
+                         config_path=THIS_DIR / 'resources/mri-config.json',
                          output_dir=fake_ds_dir, name='test_dataset')
-    seq_ids = mrd.get_sequence_ids()
-
-    for seq_id in seq_ids:
-        mrd_num_subjects = set()
-        for subject, session, run, seq in mrd.traverse_horizontal(seq_id):
-            mrd_num_subjects.add(subject)
-            assert seq['RepetitionTime'].get_value() == repetition_time
-            assert seq['EchoTrainLength'].get_value() == echo_train_length
-            assert seq['FlipAngle'].get_value() == flip_angle
-        assert set(mrd.get_subject_ids(seq_id)) == mrd_num_subjects
+    set_parameters(mrd, repetition_time, echo_train_length, flip_angle)
     shutil.rmtree(fake_ds_dir)
     return
 
 
-@settings(suppress_health_check=[HealthCheck.too_slow], max_examples=50, deadline=None)
+@settings(suppress_health_check=[HealthCheck.too_slow], max_examples=50,
+          deadline=None)
 @given(st.integers(min_value=4, max_value=10),
        st.floats(allow_nan=False,
                  allow_infinity=False),
@@ -63,8 +57,14 @@ def test_parse_dataset_no_echo_numbers(num_subjects,
                                           repetition_time,
                                           echo_train_length,
                                           flip_angle)
-    mrd = import_dataset(fake_ds_dir, config_path=THIS_DIR / 'resources/mri-config-2.json',
+    mrd = import_dataset(fake_ds_dir,
+                         config_path=THIS_DIR / 'resources/mri-config-2.json',
                          output_dir=fake_ds_dir, name='test_dataset')
+    set_parameters(mrd, repetition_time, echo_train_length, flip_angle)
+    return
+
+
+def set_parameters(mrd, repetition_time, echo_train_length, flip_angle):
     seq_ids = mrd.get_sequence_ids()
 
     for seq_id in seq_ids:
@@ -75,8 +75,6 @@ def test_parse_dataset_no_echo_numbers(num_subjects,
             assert seq['EchoTrainLength'].get_value() == echo_train_length
             assert seq['FlipAngle'].get_value() == flip_angle
         assert set(mrd.get_subject_ids(seq_id)) == mrd_num_subjects
-    shutil.rmtree(fake_ds_dir)
-    return
 
 
 def test_config_dict():
@@ -94,7 +92,8 @@ def test_empty_folders():
         filepath = subfolder / "test.dcm"
         filepath.touch()
 
-        mrd = import_dataset(folder_path, config_path=THIS_DIR / 'resources/mri-config.json',
+        mrd = import_dataset(folder_path,
+                             config_path=THIS_DIR / 'resources/mri-config.json',
                              output_dir=folder_path, name='test_dataset')
[…]
