### 2024-02-05 Open-Minds-Lab/MRdataset#24 @sinhaharsh
The installation instructions are now updated. Installing MRdataset will also install latest protocol from PyPI
