### 2024-02-06 #6269 @mstimberg
Many thanks @htwangtw for your feedback so far.
@djmannion could you give me a rough estimate when you will be able to have a look at the MRdataset software/paper? Of course, if there is anything unclear about the process, please don't hesitate to ask for help or clarifications!
