### 2024-02-06 Open-Minds-Lab/MRdataset#25 @sinhaharsh
Thanks Hao !
BIDSDataset is now included in a new PyPI release. Please try BIDS option again, it should work now.
```
pip install -U MRdataset --no-cache-dir --ignore-installed
```
