### 2024-02-06 Open-Minds-Lab/MRdataset#25 @sinhaharsh
Try the following code block. I have added configuration file, and the path to the example dataset. The same example is in the Jupyter notebook as well.

```python
import zipfile
MRdataset_root = Path(__name__).resolve().parents[1]
DATA_ARCHIVE = MRdataset_root / 'MRdataset/tests/resources/example_bids_data.zip'
DATA_ROOT = Path('/tmp/')
with zipfile.ZipFile(DATA_ARCHIVE, 'r') as zip_ref:
    zip_ref.extractall(DATA_ROOT)
print((DATA_ROOT/'example_bids_dataset').exists())
bids_dataset = import_dataset(data_source=DATA_ROOT/'example_bids_dataset',
                               ds_format='bids')
```
