### 2024-02-06 Open-Minds-Lab/MRdataset#25 @raamana
Hi @sinhaharsh , the [above](https://github.com/Open-Minds-Lab/MRdataset/issues/25#issuecomment-1930003476) might not work unless we make a new release that includes the BIDS dataset

I think Hao would have a BIDS dataset on her computer already - its better to target your instructions to a local path of a BIDS dataset
