### 2024-02-06 5f24ee8
Revise pyproject.toml : Add resources to build wheel - required for config file Correction to email address
