+++ pyproject.toml
@@ -10,7 +10,7 @@ readme = "README.rst"
 license = { file = "LICENSE" }
 requires-python = ">=3.8"
 authors = [
-    { name = "Harsh Sinha", email = "sinha.harsh@pitt.edu" },
+    { name = "Harsh Sinha", email = "harsh.sinha@pitt.edu" },
     { name = "Pradeep Raamana", email = "raamana@gmail.com" },
 ]
 keywords = [
@@ -42,6 +42,7 @@ Documentation = "https://open-minds-lab.github.io/MRdataset/"
 [tool.hatch.build]
 include = [
     "MRdataset/*.py",
+    "MRdataset/resources/*",
 ]
 exclude = [
     "MRdataset/tests/*",
