### 2024-02-08 Open-Minds-Lab/MRdataset#29 @djmannion
[JOSS] error when trying to reproduce Figure 3
https://github.com/openjournals/joss-reviews/issues/6269

* MRdataset version: 0.1.dev862+g33355da (installed via the [docs instructions](https://open-minds-lab.github.io/MRdataset/installation.html#from-sources))
* Python version: 3.11
* Operating System: Linux

### Description

Attempting to run the code shown in Figure 3 of the JOSS manuscript.

### What I Did

```python
# get an example DICOM from pydicom
import pathlib
import pydicom
demo_dcm = pathlib.Path(pydicom.data.get_testdata_file("MR_small.dcm")).parent

# from Figure 3
from MRdataset import import_dataset
ds = import_dataset(str(demo_dcm), format='dicom')
```

gives

```
dicom.py:MRdataset:__init__:80: Unable to read config file /home/damien/reviews/MRdataset/venv_repo/lib/python3.11/site-packages/MRdataset/resources/mri-config.json
Traceback (most recent call last):
  File "/home/damien/reviews/MRdataset/f3_err.py", line 9, in <module>
    ds = import_dataset(str(demo_dcm), format='dicom')
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/damien/reviews/MRdataset/venv_repo/lib/python3.11/site-packages/MRdataset/common.py", line 95, in import_dataset
    dataset = dataset_class(
              ^^^^^^^^^^^^^^
  File "/home/damien/reviews/MRdataset/venv_repo/lib/python3.11/site-packages/MRdataset/dicom.py", line 81, in __init__
    raise e
  File "/home/damien/reviews/MRdataset/venv_repo/lib/python3.11/site-packages/MRdataset/dicom.py", line 78, in __init__
    self.config_dict = read_json(Path(self.config_path))
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/damien/reviews/MRdataset/venv_repo/lib/python3.11/site-packages/MRdataset/utils.py", line 207, in read_json
    raise FileNotFoundError(f'File not found: {filepath}')
FileNotFoundError: File not found: /home/damien/reviews/MRdataset/venv_repo/lib/python3.11/site-packages/MRdataset/resources/mri-config.json
```
