### 2024-02-08 Open-Minds-Lab/MRdataset#29 @sinhaharsh
It seems that the resources folder was not included while publishing the library on PyPI. As of now you may try out the code in Figure 3 in the tutorial [notebook](https://nbviewer.org/github/Open-Minds-Lab/MRdataset/blob/main/docs/usage.ipynb).
