### 2024-02-08 Open-Minds-Lab/MRdataset#31 @djmannion
[JOSS] terms in Figure 2
https://github.com/openjournals/joss-reviews/issues/6269

In Figure 2, the 'vertical traversal' is shown as varying in 'SITE' but described as "across multiple sequences for a given subject". Maybe I am just misinterpreting, but should 'SITE' be 'SEQUENCE'?
