### 2024-02-08 Open-Minds-Lab/MRdataset#31 @sinhaharsh
Regarding Figure 2, the vertical traversal shown is across sequences, namely T1w, DWI, fMRI, T2w. The purpose of this figure is to demonstrate that each site uses a specific acquisition protocol for collecting MR scans. This protocol contains several sequences, each of which has multiple parameters. The panels in the figure represent the interface that appears on the scanner during acquisition.

It's important to note that the direction for the site should be in the Z-direction (which is into the page). As you can see from the figure, the front panels display SITE 1, while the back panels display SITE 2.

However, the direction for the site has been omitted for the sake of brevity. I hope this clarifies things for you. Please let me know if you require any additional information. I can add the direction for site if you think it would be helpful.
