### 2024-02-08 Open-Minds-Lab/MRdataset#31 @djmannion
Thanks for the explanation - I had interpreted the panels as the vertical traversal (like how the panels are the horizontal traversal), not the sequences within each panel. Maybe the 'vertical traversal' arrow could start at the top the T1w box and end at the bottom of the T2w box rather than encompassing all the panels. I'll leave it to your judgement.
