### 2024-02-09 Open-Minds-Lab/MRdataset#31 @raamana
thanks Damien - you're not wrong in interpreting that way - given we were too close to the project, we didn't see that confusion can arise. We will revise the figure to be more unambiguous and self-explanatory
