### 2024-02-09 435fc76
Modify CI GitHub Action 1) Modify working directory so that pytest uses installed version of the library (in site-packages), and not that of Github repository 2) Include tests in pyproject.toml
Related to https://github.com/Open-Minds-Lab/MRdataset/issues/29
