### 2024-02-21 24c1a8f
Add dev requirements in pyproject.toml Deprecate requirements-dev.txt
