### 2024-02-22 f1072dd
Modify BidsDataset to inform the user that datatype is not supported yet.
