### 2024-02-23 1d79b5e
Modify regex expression to exclude underscore (acts as a delimiter for entities)
Add logging message
