+++ MRdataset/bids.py
@@ -131,7 +131,11 @@ class BidsDataset(BaseDataset, ABC):
                 sequences.append(seq)
             else:
                 if name not in SUPPORTED_BIDS_DATATYPES:
-                    logger.error(f'Unsupported datatype: {name}. Skipping it.')
+                    logger.error(f'MRdataset primarily supports '
+                                 f'{SUPPORTED_BIDS_DATATYPES}'
+                                 f'It seems the parameters in '
+                                 f'this sequence are invalid or '
+                                 f'not supported yet. Skipping it.')
         return sequences
 
     @staticmethod
@@ -141,7 +145,7 @@ class BidsDataset(BaseDataset, ABC):
         Example filename : sub-01_ses-imagery01_task-imagery_run-01_bold.json
         """
         # Regular expression pattern
-        pattern = r'run-\d+'
+        pattern = r'run-[^_]+'
         # Extracting substring using regex
         match = search(pattern, str(filename))
 
