### 2024-02-23 27d7b70
Modify is_bids_file to ignore JSON with .bidsignore
