### 2024-02-27 ec88127
Modify config.py : Add supported modalities
