+++ MRdataset/dicom_utils.py
@@ -31,6 +31,8 @@ def is_bids_file(filename: Union[str, Path]):
         return False
     if 'bidsignore' in str(filename):
         return False
+    if 'sourcedata' in str(filename):
+        return False
     # Regular expression pattern
     pattern = r'sub-[^_]+'
     # Extracting substring using regex
