### 2024-02-28 #6269 @htwangtw
I am happy with the changes - the final comment is that you can omit the test data for basic installation to reduce the size of the package, and leave the data for a dev or full installation. It's up to you :)
