### 2024-03-05 73c3d2b
Modify is_bids_file
Skip physiological and other continuous recordings
