### 2024-03-09 75facee
Modify figure for vertical and horizontal traversal
