### 2024-03-09 #6269 @sinhaharsh
Thank you @mstimberg , I finished the tasks above, double checked authors and affiliations (including ORCIDs),
please kindly have the needed information:

Release version: v0.4.4

Archive DOI: [10.5281/zenodo.10798409](https://doi.org/10.5281/zenodo.10798409)

Apache 2.0 license with Github [repo](https://github.com/Open-Minds-Lab/MRdataset?tab=License-1-ov-file) and [Zenodo](https://doi.org/10.5281/zenodo.10798409)
