### 2024-03-21 #6269 @Kevin-Mattheus-Moerman
@sinhaharsh as AEiC I will now help process the final steps for acceptance in JOSS. I have checked this review, the repository, the paper, and the archive link. All seems in order so I will now proceed with acceptance in JOSS. However, I do have the below recommendations in relation to your repository (these are not required changes for JOSS, merely recommendations).

* I recommend that you edit your plain LICENSE file to confirm to the standard Apache text as I think currently this is causing GitHub not to recognise the license and render the name here properly:
![image](https://github.com/openjournals/joss-reviews/assets/8392709/c62d3c10-fe85-4be9-ba6c-ca5134879b91)
* It may be good to extend your README.md file further, e.g. extend the description slightly, and consider adding short headings/sections on Documentation, Testing, Contributing (to link to your contributing file, although I see your documentation does too, I just expect this information in a README too), and the License (linking to the license file). On the documentation and the license, you could consider adding a license and a documentation badge as many projects do, this way these are clear at the top too. Again, just suggestions to make the README feel/look more complete especially to new users seeing your project.
