### 2024-03-21 #6269 @raamana
Thank you Kevin - the old license text is likely a result of a template generator, and carried over through the years, just changed it. we will get the rest done.
